ecdsa::der

Struct Signature

pub struct Signature<C>
where
    C: EcdsaCurve,
    MaxSize<C>: ArraySize,
    <FieldBytesSize<C> as Add>::Output: Add<MaxOverhead> + ArraySize,
{
    bytes: Array<u8, MaxSize<C>>,
    r_range: Range<usize>,
    s_range: Range<usize>,
}

ASN.1 DER-encoded signature as specified in RFC5912 Section 6:

ECDSA-Sig-Value ::= SEQUENCE {
  r  INTEGER,
  s  INTEGER
}

Fields
bytes: Array<u8, MaxSize<C>>

ASN.1 DER-encoded signature data

r_range: Range<usize>

Range of the r value within the signature

s_range: Range<usize>

Range of the s value within the signature

Implementations

impl<C> Signature<C>
where C: EcdsaCurve, MaxSize<C>: ArraySize, <FieldBytesSize<C> as Add>::Output: Add<MaxOverhead> + ArraySize,

pub fn from_bytes(input: &[u8]) -> Result<Self>

Parse signature from DER-encoded bytes.

Errors

Returns Error if input failed to parse as an ASN.1 DER-encoded ECDSA signature.

pub(crate) fn from_components(r: &[u8], s: &[u8]) -> Result<Self>

Create an ASN.1 DER encoded signature from big endian r and s scalar components.

pub fn as_bytes(&self) -> &[u8]

Borrow this signature as a byte slice

pub fn to_bytes(&self) -> Box<[u8]>

Serialize this signature as a boxed byte slice

pub fn len(&self) -> usize

Get the length of the signature in bytes

pub(crate) fn r(&self) -> &[u8]

Get the r component of the signature (leading zeros removed)

pub(crate) fn s(&self) -> &[u8]

Get the s component of the signature (leading zeros removed)