Module ecdsa::hazmat

Available on crate feature hazmat only.

Low-level ECDSA primitives.

⚠️ Warning: Hazmat!

YOU PROBABLY DON’T WANT TO USE THESE!

These primitives are easy-to-misuse low-level interfaces.

If you are an end user / non-expert in cryptography, do not use these! Failure to use them correctly can lead to catastrophic failures including FULL PRIVATE KEY RECOVERY!

Traits

DigestPrimitivedigest

Bind a preferred Digest algorithm to an elliptic curve type.

SignPrimitivearithmetic

Try to sign the given prehashed message using ECDSA.

VerifyPrimitivearithmetic

Verify the given prehashed message using ECDSA.