ecdsa::hazmat

Function bytes2scalar

pub(crate) fn bytes2scalar<C: EcdsaCurve + CurveArithmetic>(
    bytes: &[u8],
) -> Scalar<C>

Convert the provided bytestring into a Scalar for the given curve, interpreting it as big endian, zero-padding or truncating it to the bit length of n (curve order) if necessary, and then reducing it mod n.