Trait ecdsa::hazmat::DigestPrimitive

pub trait DigestPrimitive: PrimeCurve {
    type Digest: BlockSizeUser + Digest + FixedOutput<OutputSize = FieldBytesSize<Self>> + FixedOutputReset;
}

Bind a preferred Digest algorithm to an elliptic curve type.

Generally there is a preferred variety of the SHA-2 family used with ECDSA for a particular elliptic curve.

This trait can be used to specify it, and with it receive a blanket impl of PrehashSignature, used by signature_derive) for the Signature type for a particular elliptic curve.

Required Associated Types

type Digest: BlockSizeUser + Digest + FixedOutput<OutputSize = FieldBytesSize<Self>> + FixedOutputReset

Preferred digest to use when computing ECDSA signatures for this elliptic curve. This is typically a member of the SHA-2 family.

Object Safety

This trait is not object safe.