ecdsa::signing

Struct SigningKey

pub struct SigningKey<C>
where
    C: EcdsaCurve + CurveArithmetic,
{
    secret_scalar: NonZeroScalar<C>,
    verifying_key: VerifyingKey<C>,
}

ECDSA secret key used for signing.

Generic over prime order elliptic curves (e.g. NIST P-curves). Requires an elliptic_curve::CurveArithmetic impl on the curve.

Usage

The signature crate defines the following traits which are the primary API for signing:

Signer: sign a message using this key

DigestSigner: sign the output of a digest using this key

PrehashSigner: sign the low-level raw output bytes of a message digest

See the p256 crate for examples of using this type with a concrete elliptic curve.

Fields
secret_scalar: NonZeroScalar<C>

ECDSA signing keys are non-zero elements of a given curve’s scalar field.

verifying_key: VerifyingKey<C>

Verifying key which corresponds to this signing key.

Implementations

impl<C> SigningKey<C>
where C: EcdsaCurve + CurveArithmetic + DigestAlgorithm, Scalar<C>: Invert<Output = CtOption<Scalar<C>>>,

pub fn sign_prehash_recoverable_with_rng<R: TryCryptoRng + ?Sized>( &self, rng: &mut R, prehash: &[u8], ) -> Result<(Signature<C>, RecoveryId), R::Error>

Sign the given message prehash, using the given rng for the RFC6979 Section 3.6 “additional data”, returning a signature and recovery ID.

Errors

pub fn sign_prehash_recoverable( &self, prehash: &[u8], ) -> (Signature<C>, RecoveryId)

Sign the given message prehash, returning a signature and recovery ID.

pub fn sign_digest_recoverable<D: Digest>( &self, msg_digest: D, ) -> (Signature<C>, RecoveryId)

Sign the given message digest, returning a signature and recovery ID.

pub fn sign_recoverable(&self, msg: &[u8]) -> (Signature<C>, RecoveryId)

Sign the given message, hashing it with the curve’s default digest function, and returning a signature and recovery ID.

impl<C> SigningKey<C>
where C: EcdsaCurve + CurveArithmetic,

pub fn from_bytes(bytes: &FieldBytes<C>) -> Result<Self>

Initialize signing key from a raw scalar serialized as a byte array.

Errors

Returns an error if bytes is not a valid element of the scalar field for C.

pub fn from_slice(bytes: &[u8]) -> Result<Self>

Initialize signing key from a raw scalar serialized as a byte slice.

Errors

Returns an error if bytes is not the length of FieldBytes for C, or if it is not a valid element of the scalar field for C.

pub fn to_bytes(&self) -> FieldBytes<C>

Serialize this SigningKey as bytes

pub fn as_nonzero_scalar(&self) -> &NonZeroScalar<C>

Borrow the secret NonZeroScalar value for this key.

Security Warning
This value is key material. Please treat it with the care it deserves!

pub fn verifying_key(&self) -> &VerifyingKey<C>

Get the VerifyingKey which corresponds to this SigningKey.

pub fn random<R: CryptoRng + ?Sized>(rng: &mut R) -> Self

👎Deprecated since 0.17.0:
use the Generate trait instead

DEPRECATED: Generate a cryptographically random SigningKey.