ecdsa

Struct Signature

pub struct Signature<C: EcdsaCurve> {
    pub(crate) r: ScalarValue<C>,
    pub(crate) s: ScalarValue<C>,
}

ECDSA signature (fixed-size, a.k.a. IEEE P1363). Generic over elliptic curve types.

Serialized as fixed-sized big endian scalar values with no added framing:

r: field element size for the given curve, big-endian

s: field element size for the given curve, big-endian

Both r and s MUST be non-zero.

For example, in a curve with a 256-bit modulus like NIST P-256 or secp256k1, r and s are both 32-bytes and serialized as big endian, resulting in a signature with a total of 64-bytes.

ASN.1 DER-encoded signatures also supported via the Signature::from_der and Signature::to_der methods.

serde support

When the serde feature of this crate is enabled, it provides support for serializing and deserializing ECDSA signatures using the Serialize and Deserialize traits.

The serialization uses a hexadecimal encoding when used with “human readable” text formats, and a binary encoding otherwise.

Fields
r: ScalarValue<C>s: ScalarValue<C>
Implementations

impl<C> Signature<C>
where C: EcdsaCurve,

pub fn from_bytes(bytes: &SignatureBytes<C>) -> Result<Self>

Parse a signature from fixed-width bytes, i.e. 2 * the size of FieldBytes for a particular curve.

Errors

If the r and/or s component of the signature is out-of-range when interpreted as a big endian integer.

pub fn from_slice(slice: &[u8]) -> Result<Self>

Parse a signature from a byte slice.

Errors

Returns Error in the event the signature is not the expected size, i.e. 2 * the size of FieldBytes for a particular curve.

pub fn from_der(bytes: &[u8]) -> Result<Self>
where MaxSize<C>: ArraySize, <FieldBytesSize<C> as Add>::Output: Add<MaxOverhead> + ArraySize,

Parse a signature from ASN.1 DER.

Errors

Returns Error if input failed to parse as an ASN.1 DER-encoded ECDSA signature.

pub fn from_scalars( r: impl Into<FieldBytes<C>>, s: impl Into<FieldBytes<C>>, ) -> Result<Self>

Create a Signature from the serialized r and s scalar values which comprise the signature.

Errors

If the r and/or s component of the signature is out-of-range when interpreted as a big endian integer.

pub fn split_bytes(&self) -> (FieldBytes<C>, FieldBytes<C>)

Split the signature into its r and s components, represented as bytes.

pub fn to_bytes(&self) -> SignatureBytes<C>

Serialize this signature as bytes.

pub fn to_der(&self) -> Signature<C>
where MaxSize<C>: ArraySize, <FieldBytesSize<C> as Add>::Output: Add<MaxOverhead> + ArraySize,

Serialize this signature as ASN.1 DER.

pub fn to_vec(&self) -> Vec<u8>

Convert this signature into a byte vector.

impl<C> Signature<C>
where C: EcdsaCurve + CurveArithmetic,

pub fn r(&self) -> NonZeroScalar<C>

Get the r component of this signature

pub fn s(&self) -> NonZeroScalar<C>

Get the s component of this signature

pub fn split_scalars(&self) -> (NonZeroScalar<C>, NonZeroScalar<C>)

Split the signature into its r and s scalars.

pub fn normalize_s(&self) -> Self

Normalize signature into “low S” form described in BIP 0062: Dealing with Malleability.