ecdsa
Struct SignatureWithOid

pub struct SignatureWithOid<C: PrimeCurve> { /* private fields */ }

Available on crate feature digest only.

An extended Signature type which is parameterized by an ObjectIdentifier which identifies the ECDSA variant used by a particular signature.

Valid ObjectIdentifiers are defined in RFC5758  3.2:

SHA-224: ECDSA_SHA224_OID (1.2.840.10045.4.3.1)

SHA-256: ECDSA_SHA256_OID (1.2.840.10045.4.3.2)

SHA-384: ECDSA_SHA384_OID (1.2.840.10045.4.3.3)

SHA-512: ECDSA_SHA512_OID (1.2.840.10045.4.3.4)

Implementations

impl<C> SignatureWithOid<C>
where C: PrimeCurve,

pub fn new(signature: Signature<C>, oid: ObjectIdentifier) -> Result<Self>

Create a new signature with an explicitly provided OID.

OID must begin with 1.2.840.10045.4, the RFC5758 OID prefix for ECDSA variants.

pub fn new_with_digest<D>(signature: Signature<C>) -> Result<Self>
where D: AssociatedOid + Digest,

Create a new signature, determining the OID from the given digest.

Supports SHA-2 family digests as enumerated in RFC5758  3.2, i.e. SHA-224, SHA-256, SHA-384, or SHA-512.

pub fn from_bytes_with_digest<D>(bytes: &SignatureBytes<C>) -> Result<Self>
where D: AssociatedOid + Digest, SignatureSize<C>: ArrayLength<u8>,

Parse a signature from fixed-with bytes.

pub fn from_slice_with_digest<D>(slice: &[u8]) -> Result<Self>
where D: AssociatedOid + Digest, SignatureSize<C>: ArrayLength<u8>,

Parse a signature from a byte slice.

pub fn signature(&self) -> &Signature<C>

Get the fixed-width ECDSA signature.

pub fn oid(&self) -> ObjectIdentifier

Get the ECDSA OID for this signature.

pub fn to_bytes(&self) -> SignatureBytes<C>
where SignatureSize<C>: ArrayLength<u8>,

Serialize this signature as bytes.