Trait ecdsa::NormalizeLow

pub trait NormalizeLow: Sized {
    // Required method
    fn normalize_low(&self) -> (Self, bool);
}

Normalize a scalar (i.e. ECDSA S) to the lower half the field, as described in BIP 0062: Dealing with Malleability.

Required Methods

fn normalize_low(&self) -> (Self, bool)

Normalize scalar to the lower half of the field (i.e. negate it if it’s larger than half the curve’s order). Returns a tuple with the new scalar and a boolean indicating whether the given scalar was in the higher half.

May be implemented to work in variable time.

Object Safety

This trait is not object safe.