ecdsa
Module hazmat

Low-level ECDSA primitives.

⚠️ Warning: Hazmat!

YOU PROBABLY DON’T WANT TO USE THESE!

These primitives are easy-to-misuse low-level interfaces intended to be implemented by elliptic curve crates and consumed only by this crate!

If you are an end user / non-expert in cryptography, do not use these! Failure to use them correctly can lead to catastrophic failures including FULL PRIVATE KEY RECOVERY!

Traits

DigestPrimitive

Bind a preferred Digest algorithm to an elliptic curve type.

FromDigest

Instantiate this type from the output of a digest.

RecoverableSignPrimitive

SignPrimitive for signature implementations that can provide public key recovery implementation.

SignPrimitive

Try to sign the given prehashed message using ECDSA.

VerifyPrimitive

Verify the given prehashed message using ECDSA.