Trait ecdsa::hazmat::RecoverableSignPrimitive

pub trait RecoverableSignPrimitive<C>where
    C: Curve + ProjectiveArithmetic,
    SignatureSize<C>: ArrayLength<u8>,{
    // Required method
    fn try_sign_recoverable_prehashed<K: Borrow<Scalar<C>> + Invert<Output = Scalar<C>>>(
        &self,
        ephemeral_scalar: &K,
        hashed_msg: &Scalar<C>
    ) -> Result<(Signature<C>, bool)>;
}

SignPrimitive for signature implementations that can provide public key recovery implementation.

Required Methods

fn try_sign_recoverable_prehashed<K: Borrow<Scalar<C>> + Invert<Output = Scalar<C>>>( &self, ephemeral_scalar: &K, hashed_msg: &Scalar<C> ) -> Result<(Signature<C>, bool)>

Try to sign the prehashed message.

Accepts the same arguments as SignPrimitive::try_sign_prehashed but returns a boolean flag which indicates whether or not the y-coordinate of the computed ğ‘ = ğ‘˜Ã—ğ‘® point is odd, which can be incorporated into recoverable signatures.