ecdsa::hazmat
Trait SignPrimitive

pub trait SignPrimitive<C>
where
    C: Curve + ProjectiveArithmetic,
    SignatureSize<C>: ArrayLength<u8>,
{
    // Required method
    fn try_sign_prehashed<K: Borrow<Scalar<C>> + Invert<Output = Scalar<C>>>(
        &self,
        ephemeral_scalar: &K,
        hashed_msg: &Scalar<C>,
    ) -> Result<Signature<C>>;
}

Try to sign the given prehashed message using ECDSA.

This trait is intended to be implemented on a type with access to the secret scalar via &self, such as particular curve’s Scalar type, or potentially a key handle to a hardware device.

Required Methods

fn try_sign_prehashed<K: Borrow<Scalar<C>> + Invert<Output = Scalar<C>>>( &self, ephemeral_scalar: &K, hashed_msg: &Scalar<C>, ) -> Result<Signature<C>>

Try to sign the prehashed message.

Accepts the following arguments:

ephemeral_scalar: ECDSA k value. MUST BE UNIFORMLY RANDOM!!!

hashed_msg: scalar computed from a hashed message digest to be signed. MUST BE OUTPUT OF A CRYPTOGRAPHICALLY SECURE DIGEST ALGORITHM!!!

Object Safety

This trait is not object safe.