Trait ecdsa::hazmat::VerifyPrimitive

pub trait VerifyPrimitive<C>: AffineCoordinates<FieldRepr = FieldBytes<C>> + Copy + Sized
where
    C: PrimeCurve + CurveArithmetic<AffinePoint = Self>,
    SignatureSize<C>: ArrayLength<u8>,
{
    // Provided methods
    fn verify_prehashed(
        &self,
        z: &FieldBytes<C>,
        sig: &Signature<C>,
    ) -> Result<()> { ... }
    fn verify_digest<D>(&self, msg_digest: D, sig: &Signature<C>) -> Result<()>
       where D: FixedOutput<OutputSize = FieldBytesSize<C>> { ... }
}

Verify the given prehashed message using ECDSA.

This trait is intended to be implemented on type which can access the affine point represeting the public key via &self, such as a particular curve’s AffinePoint type.

Provided Methods

fn verify_prehashed(&self, z: &FieldBytes<C>, sig: &Signature<C>) -> Result<()>

Verify the prehashed message against the provided ECDSA signature.

Accepts the following arguments:

z: message digest to be verified. MUST BE OUTPUT OF A CRYPTOGRAPHICALLY SECURE DIGEST ALGORITHM!!!

sig: signature to be verified against the key and message

fn verify_digest<D>(&self, msg_digest: D, sig: &Signature<C>) -> Result<()>
where D: FixedOutput<OutputSize = FieldBytesSize<C>>,

Verify message digest against the provided signature.

Object Safety

This trait is not object safe.