Trait ecdsa::hazmat::VerifyPrimitive

pub trait VerifyPrimitive<C>: AffineXCoordinate<C> + Copy + Sized where
    C: PrimeCurve + AffineArithmetic<AffinePoint = Self> + ProjectiveArithmetic,
    Scalar<C>: Reduce<C::UInt>,
    SignatureSize<C>: ArrayLength<u8>, {
    fn verify_prehashed(&self, z: Scalar<C>, sig: &Signature<C>) -> Result<()> { ... }
}

Verify the given prehashed message using ECDSA.

This trait is intended to be implemented on type which can access the affine point represeting the public key via &self, such as a particular curve’s AffinePoint type.

Provided methods

fn verify_prehashed(&self, z: Scalar<C>, sig: &Signature<C>) -> Result<()>

Verify the prehashed message against the provided signature

Accepts the following arguments:

z: prehashed message to be verified

sig: signature to be verified against the key and message