ecdsa::hazmat
Trait VerifyPrimitive

pub trait VerifyPrimitive<C>
where
    C: Curve + ProjectiveArithmetic,
    SignatureSize<C>: ArrayLength<u8>,
{
    // Required method
    fn verify_prehashed(
        &self,
        hashed_msg: &Scalar<C>,
        signature: &Signature<C>,
    ) -> Result<()>;
}

Verify the given prehashed message using ECDSA.

This trait is intended to be implemented on type which can access the affine point represeting the public key via &self, such as a particular curve’s AffinePoint type.

Required Methods

fn verify_prehashed( &self, hashed_msg: &Scalar<C>, signature: &Signature<C>, ) -> Result<()>

Verify the prehashed message against the provided signature

Accepts the following arguments:

hashed_msg: prehashed message to be verified

signature: signature to be verified against the key and message