Struct ecdsa::EncodedPoint

pub struct EncodedPoint<C>where
    C: Curve,
    <<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output: Add<UInt<UTerm, B1>> + ArrayLength<u8>,
    <<<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output as Add<UInt<UTerm, B1>>>::Output: ArrayLength<u8>,{ /* private fields */ }

SEC1 encoded curve point.

This type is an enum over the compressed and uncompressed encodings, useful for cases where either encoding can be supported, or conversions between the two forms.

Implementations

impl<C> EncodedPoint<C>where C: Curve, <<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output: Add<UInt<UTerm, B1>> + ArrayLength<u8>, <<<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output as Add<UInt<UTerm, B1>>>::Output: ArrayLength<u8>,

pub fn from_bytes(input: impl AsRef<[u8]>) -> Result<EncodedPoint<C>, Error>

Decode elliptic curve point (compressed or uncompressed) from the Elliptic-Curve-Point-to-Octet-String encoding described in SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.3 (page 10).

http://www.secg.org/sec1-v2.pdf

pub fn from_untagged_bytes( bytes: &GenericArray<u8, <<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output> ) -> EncodedPoint<C>

Decode elliptic curve point from raw uncompressed coordinates, i.e. encoded as the concatenated x || y coordinates with no leading SEC1 tag byte (which would otherwise be 0x04 for an uncompressed point).

pub fn from_affine_coordinates( x: &GenericArray<u8, <<C as Curve>::UInt as ArrayEncoding>::ByteSize>, y: &GenericArray<u8, <<C as Curve>::UInt as ArrayEncoding>::ByteSize>, compress: bool ) -> EncodedPoint<C>

Encode an elliptic curve point from big endian serialized coordinates (with optional point compression)

pub fn from_secret_key( secret_key: &SecretKey<C>, compress: bool ) -> EncodedPoint<C>where C: Curve + ProjectiveArithmetic, <C as AffineArithmetic>::AffinePoint: ToEncodedPoint<C>, <C as ScalarArithmetic>::Scalar: Zeroize,

Compute EncodedPoint representing the public key for the provided SecretKey.

The compress flag requests point compression.

pub fn identity() -> EncodedPoint<C>

Return EncodedPoint representing the additive identity (a.k.a. point at infinity)

pub fn len(&self) -> usize

Get the length of the encoded point in bytes

pub fn as_bytes(&self) -> &[u8]

Get byte slice containing the serialized EncodedPoint.

pub fn to_untagged_bytes( &self ) -> Option<GenericArray<u8, <<<C as Curve>::UInt as ArrayEncoding>::ByteSize as Add<<<C as Curve>::UInt as ArrayEncoding>::ByteSize>>::Output>>where C: Curve + ProjectiveArithmetic, <C as AffineArithmetic>::AffinePoint: DecompressPoint<C> + ToEncodedPoint<C>,

Serialize point as raw uncompressed coordinates without tag byte, i.e. encoded as the concatenated x || y coordinates.

pub fn is_compact(&self) -> bool

Is this EncodedPoint compact?

pub fn is_compressed(&self) -> bool

Is this EncodedPoint compressed?

pub fn is_identity(&self) -> bool

Is this EncodedPoint the additive identity? (a.k.a. point at infinity)

pub fn compress(&self) -> EncodedPoint<C>

Compress this EncodedPoint, returning a new EncodedPoint.

pub fn decompress(&self) -> Option<EncodedPoint<C>>where C: Curve + ProjectiveArithmetic, <C as AffineArithmetic>::AffinePoint: DecompressPoint<C> + ToEncodedPoint<C>,

Decompress this EncodedPoint, returning a new EncodedPoint.

pub fn encode<T>(encodable: T, compress: bool) -> EncodedPoint<C>where T: ToEncodedPoint<C>,

Encode an EncodedPoint from the desired type

pub fn decode<T>(&self) -> Result<T, Error>where T: FromEncodedPoint<C>,

Decode this EncodedPoint into the desired type

pub fn tag(&self) -> Tag

Get the SEC1 tag for this EncodedPoint

pub fn coordinates(&self) -> Coordinates<'_, C>

Get the Coordinates for this EncodedPoint.

pub fn x( &self ) -> Option<&GenericArray<u8, <<C as Curve>::UInt as ArrayEncoding>::ByteSize>>

Get the x-coordinate for this EncodedPoint.

Returns None if this point is the identity point.

pub fn y( &self ) -> Option<&GenericArray<u8, <<C as Curve>::UInt as ArrayEncoding>::ByteSize>>

Get the y-coordinate for this EncodedPoint.

Returns None if this point is compressed or the identity point.