Struct ecdsa::Signature

pub struct Signature<C: Curve> where
    SignatureSize<C>: ArrayLength<u8>,  { /* private fields */ }

ECDSA signature (fixed-size). Generic over elliptic curve types.

Serialized as fixed-sized big endian scalar values with no added framing:

r: field element size for the given curve, big-endian

s: field element size for the given curve, big-endian

For example, in a curve with a 256-bit modulus like NIST P-256 or secp256k1, r and s will both be 32-bytes, resulting in a signature with a total of 64-bytes.

ASN.1 DER-encoded signatures also supported via the Signature::from_der and Signature::to_der methods.

Implementations

impl<C> Signature<C> where
C: Curve,
SignatureSize<C>: ArrayLength<u8>,

pub fn from_scalars(
r: impl Into<FieldBytes<C>>,
s: impl Into<FieldBytes<C>>
) -> Result<Self>

Create a Signature from the serialized r and s scalar values which comprise the signature.

pub fn from_der(bytes: &[u8]) -> Result<Self> where
MaxSize<C>: ArrayLength<u8>,
<FieldSize<C> as Add>::Output: Add<MaxOverhead> + ArrayLength<u8>,

Parse a signature from ASN.1 DER

pub fn to_der(&self) -> Signature<C> where
MaxSize<C>: ArrayLength<u8>,
<FieldSize<C> as Add>::Output: Add<MaxOverhead> + ArrayLength<u8>,

Serialize this signature as ASN.1 DER

impl<C> Signature<C> where
C: Curve + ProjectiveArithmetic,
SignatureSize<C>: ArrayLength<u8>,

pub fn r(&self) -> NonZeroScalar<C>

Get the r component of this signature

pub fn s(&self) -> NonZeroScalar<C>

Get the s component of this signature

pub fn normalize_s(&mut self) -> Result<bool> where
Scalar<C>: NormalizeLow,

Normalize signature into “low S” form as described in BIP 0062: Dealing with Malleability.