Struct ecdsa::Signature

pub struct Signature<C: PrimeCurve> where
    SignatureSize<C>: ArrayLength<u8>,  { /* fields omitted */ }

ECDSA signature (fixed-size). Generic over elliptic curve types.

Serialized as fixed-sized big endian scalar values with no added framing:

r: field element size for the given curve, big-endian

s: field element size for the given curve, big-endian

For example, in a curve with a 256-bit modulus like NIST P-256 or secp256k1, r and s will both be 32-bytes, resulting in a signature with a total of 64-bytes.

ASN.1 DER-encoded signatures also supported via the Signature::from_der and Signature::to_der methods.

serde support

When the serde feature of this crate is enabled, it provides support for serializing and deserializing ECDSA signatures using the Serialize and Deserialize traits.

The serialization uses a 64-byte fixed encoding when used with binary formats, and a hexadecimal encoding when used with text formats.

Implementations

impl<C> Signature<C> where
C: PrimeCurve,
SignatureSize<C>: ArrayLength<u8>,

pub fn from_der(bytes: &[u8]) -> Result<Self> where
MaxSize<C>: ArrayLength<u8>,
<FieldSize<C> as Add>::Output: Add<MaxOverhead> + ArrayLength<u8>,

Parse a signature from ASN.1 DER

pub fn from_scalars(
r: impl Into<FieldBytes<C>>,
s: impl Into<FieldBytes<C>>
) -> Result<Self>

Create a Signature from the serialized r and s scalar values which comprise the signature.

pub fn split_bytes(&self) -> (FieldBytes<C>, FieldBytes<C>)

Split the signature into its r and s components, represented as bytes.

pub fn to_der(&self) -> Signature<C> where
MaxSize<C>: ArrayLength<u8>,
<FieldSize<C> as Add>::Output: Add<MaxOverhead> + ArrayLength<u8>,

Serialize this signature as ASN.1 DER

pub fn to_vec(&self) -> Vec<u8>

Convert this signature into a byte vector.