Struct ecdsa::SigningKey

pub struct SigningKey<C>
where
    C: PrimeCurve + CurveArithmetic,
    Scalar<C>: Invert<Output = CtOption<Scalar<C>>> + SignPrimitive<C>,
    SignatureSize<C>: ArrayLength<u8>,
{ /* private fields */ }

ECDSA secret key used for signing. Generic over prime order elliptic curves (e.g. NIST P-curves)

Requires an elliptic_curve::CurveArithmetic impl on the curve, and a SignPrimitive impl on its associated Scalar type.

Usage

The signature crate defines the following traits which are the primary API for signing:

Signer: sign a message using this key

DigestSigner: sign the output of a Digest using this key

PrehashSigner: sign the low-level raw output bytes of a message digest

See the p256 crate for examples of using this type with a concrete elliptic curve.

Implementations

impl<C> SigningKey<C>
where C: PrimeCurve + CurveArithmetic + DigestPrimitive, Scalar<C>: Invert<Output = CtOption<Scalar<C>>> + SignPrimitive<C>, SignatureSize<C>: ArrayLength<u8>,

pub fn sign_prehash_recoverable( &self, prehash: &[u8], ) -> Result<(Signature<C>, RecoveryId)>

Sign the given message prehash, returning a signature and recovery ID.

pub fn sign_digest_recoverable<D>( &self, msg_digest: D, ) -> Result<(Signature<C>, RecoveryId)>
where D: Digest,

Sign the given message digest, returning a signature and recovery ID.

pub fn sign_recoverable(&self, msg: &[u8]) -> Result<(Signature<C>, RecoveryId)>

Sign the given message, hashing it with the curve’s default digest function, and returning a signature and recovery ID.

impl<C> SigningKey<C>
where C: PrimeCurve + CurveArithmetic, Scalar<C>: Invert<Output = CtOption<Scalar<C>>> + SignPrimitive<C>, SignatureSize<C>: ArrayLength<u8>,

pub fn random(rng: &mut impl CryptoRngCore) -> Self

Generate a cryptographically random SigningKey.

pub fn from_bytes(bytes: &FieldBytes<C>) -> Result<Self>

Initialize signing key from a raw scalar serialized as a byte array.

pub fn from_slice(bytes: &[u8]) -> Result<Self>

Initialize signing key from a raw scalar serialized as a byte slice.

pub fn to_bytes(&self) -> FieldBytes<C>

Serialize this SigningKey as bytes

pub fn as_nonzero_scalar(&self) -> &NonZeroScalar<C>

Borrow the secret NonZeroScalar value for this key.

⚠️ Warning

This value is key material.

Please treat it with the care it deserves!

pub fn verifying_key(&self) -> &VerifyingKey<C>

Get the VerifyingKey which corresponds to this SigningKey.