Struct ecdsa::VerifyingKey

pub struct VerifyingKey<C>
where
    C: PrimeCurve + CurveArithmetic,
{ /* private fields */ }

ECDSA public key used for verifying signatures. Generic over prime order elliptic curves (e.g. NIST P-curves)

Requires an elliptic_curve::CurveArithmetic impl on the curve, and a VerifyPrimitive impl on its associated AffinePoint type.

Usage

The signature crate defines the following traits which are the primary API for verifying:

Verifier: verify a message against a provided key and signature

DigestVerifier: verify a message Digest against a provided key and signature

PrehashVerifier: verify the low-level raw output bytes of a message digest

See the p256 crate for examples of using this type with a concrete elliptic curve.

serde support

When the serde feature of this crate is enabled, it provides support for serializing and deserializing ECDSA signatures using the Serialize and Deserialize traits.

The serialization leverages the encoding used by the PublicKey type, which is a binary-oriented ASN.1 DER encoding.

Implementations

impl<C> VerifyingKey<C>
where C: PrimeCurve + CurveArithmetic, AffinePoint<C>: DecompressPoint<C> + FromEncodedPoint<C> + ToEncodedPoint<C> + VerifyPrimitive<C>, FieldBytesSize<C>: ModulusSize, SignatureSize<C>: ArrayLength<u8>,

pub fn recover_from_msg( msg: &[u8], signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>
where C: DigestPrimitive,

Recover a VerifyingKey from the given message, signature, and RecoveryId.

The message is first hashed using this curve’s DigestPrimitive.

pub fn recover_from_digest<D>( msg_digest: D, signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>
where D: Digest,

Recover a VerifyingKey from the given message Digest, signature, and RecoveryId.

pub fn recover_from_prehash( prehash: &[u8], signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>

Recover a VerifyingKey from the given prehash of a message, the signature over that prehashed message, and a RecoveryId.

impl<C> VerifyingKey<C>
where C: PrimeCurve + CurveArithmetic, AffinePoint<C>: FromEncodedPoint<C> + ToEncodedPoint<C>, FieldBytesSize<C>: ModulusSize,

pub fn from_sec1_bytes(bytes: &[u8]) -> Result<Self>

Initialize VerifyingKey from a SEC1-encoded public key.

pub fn from_affine(affine: AffinePoint<C>) -> Result<Self>

Initialize VerifyingKey from an affine point.

Returns an Error if the given affine point is the additive identity (a.k.a. point at infinity).

pub fn from_encoded_point(public_key: &EncodedPoint<C>) -> Result<Self>

Initialize VerifyingKey from an EncodedPoint.

pub fn to_encoded_point(&self, compress: bool) -> EncodedPoint<C>

Serialize this VerifyingKey as a SEC1 EncodedPoint, optionally applying point compression.

pub fn to_sec1_bytes(&self) -> Box<[u8]>
where C: PointCompression,

Convert this VerifyingKey into the Elliptic-Curve-Point-to-Octet-String encoding described in SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.3 (page 10).

http://www.secg.org/sec1-v2.pdf

pub fn as_affine(&self) -> &AffinePoint<C>

Borrow the inner AffinePoint for this public key.