ecdsa

Struct VerifyingKey

pub struct VerifyingKey<C>
where
    C: EcdsaCurve + CurveArithmetic,
{
    pub(crate) inner: PublicKey<C>,
}

ECDSA public key used for verifying signatures. Generic over prime order elliptic curves (e.g. NIST P-curves).

Requires an elliptic_curve::CurveArithmetic impl on the curve.

Usage

The signature crate defines the following traits which are the primary API for verifying:

Verifier: verify a message against a provided key and signature

DigestVerifier: verify a message digest against a provided key and signature

PrehashVerifier: verify the low-level raw output bytes of a message digest

See the p256 crate for examples of using this type with a concrete elliptic curve.

serde support

When the serde feature of this crate is enabled, it provides support for serializing and deserializing ECDSA signatures using the Serialize and Deserialize traits.

The serialization leverages the encoding used by the PublicKey type, which is a binary-oriented ASN.1 DER encoding.

Fields
inner: PublicKey<C>
Implementations

impl<C> VerifyingKey<C>
where C: EcdsaCurve + CurveArithmetic, AffinePoint<C>: DecompressPoint<C> + FromSec1Point<C> + ToSec1Point<C>, FieldBytesSize<C>: ModulusSize,

pub fn recover_from_msg( msg: &[u8], signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>
where C: DigestAlgorithm,

Recover a VerifyingKey from the given message, signature, and RecoveryId.

The message is first hashed using this curve’s DigestAlgorithm.

Errors

Returns Error if the recovered elliptic curve point is the additive identity.

pub fn recover_from_digest<D>( msg_digest: D, signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>
where D: Digest,

Recover a VerifyingKey from the given message Digest, signature, and RecoveryId.

Errors

Returns Error if the recovered elliptic curve point is the additive identity.

pub fn recover_from_prehash( prehash: &[u8], signature: &Signature<C>, recovery_id: RecoveryId, ) -> Result<Self>

Recover a VerifyingKey from the given prehash of a message, the signature over that prehashed message, and a RecoveryId.

Security Warning
The prehash argument must be the output of a secure digest function, e.g. Keccak256 or SHA-256.

Failure to use such a digest algorithm to compute prehash allows an attacker to solve for it in a system of linear equations that can cause the recovery function to output any public key the attacker wants.

Errors

Returns Error if the recovered elliptic curve point is the additive identity.

impl<C> VerifyingKey<C>
where C: EcdsaCurve + CurveArithmetic, AffinePoint<C>: FromSec1Point<C> + ToSec1Point<C>, FieldBytesSize<C>: ModulusSize,

pub fn from_sec1_bytes(bytes: &[u8]) -> Result<Self>

Initialize VerifyingKey from a SEC1-encoded public key.

Errors

Returns Error if bytes is not a valid SEC1 encoding of a compressed or uncompressed curve point (i.e. it should begin with 0x02, 0x03, or 0x04), or if VerifyingKey::from_sec1_point returns an error.

pub fn from_affine(affine: AffinePoint<C>) -> Result<Self>

Initialize VerifyingKey from an affine point.

Errors

Returns Error if the affine point is the additive identity (a.k.a. point at infinity).

pub fn from_sec1_point(public_key: &Sec1Point<C>) -> Result<Self>

Initialize VerifyingKey from an Sec1Point.

Errors

Returns Error if public_key does not represent a valid elliptic curve point for C.

pub fn to_sec1_point(&self, compress: bool) -> Sec1Point<C>

Serialize this VerifyingKey as a SEC1 Sec1Point, optionally applying point compression.

pub fn to_sec1_bytes(&self) -> Box<[u8]>
where C: PointCompression,

Convert this VerifyingKey into the Elliptic-Curve-Point-to-Octet-String encoding described in SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.3 (page 10).

http://www.secg.org/sec1-v2.pdf

pub fn as_affine(&self) -> &AffinePoint<C>

Borrow the inner AffinePoint for this public key.